+++ src/main/java/be/ugent/rml/functions/DynamicMultipleRecordsFunctionExecutor.java
@@ -81,24 +81,6 @@ public class DynamicMultipleRecordsFunctionExecutor implements MultipleRecordsFu
             throw new Exception("No function was defined for parameters: " + mergedArgs.keySet());
         } else {
             FunctionModel function = functionLoader.getFunction(fnTerms.get(0));
-            String param_rep_b = "http://users.ugent.be/~bjdmeest/function/grel.ttl#param_rep_b";
-            if (mergedArgs.containsKey(param_rep_b)) {
-                try {
-                    List<Object> boolList = mergedArgs.get(param_rep_b).stream().map(e -> {
-                        if(e instanceof Boolean){
-                            return e;
-                        } else if(e instanceof String){
-                            return Boolean.parseBoolean((String) e);
-                        } else {
-                            throw new IllegalArgumentException(String.format("Unable to convert %s to boolean in param_rep_b.", e));
-                        }
-                    }).collect(Collectors.toList());
-                    mergedArgs.replace(param_rep_b, boolList);
-                    logger.warn("Param_rep_b has been cast to list of booleans.");
-                } catch (Exception e) {
-                    throw new Exception("Param_rep_b couldn't be cast to a list of booleans.");
-                }
-            }
             return function.execute((Map) mergedArgs);
         }
     }